# docs/api-reference/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Bodhveda Notifications API",
        "version": "1.0",
        "description": "API for sending notifications to recipients or broadcasting to all."
    },
    "paths": {
...
        "/recipients/{recipient_id}/preferences": {
            "get": {
                "summary": "List preferences for a recipient",
                "operationId": "listRecipientPreferences",
                "tags": [
                    "Preferences"
                ],
                "description": "Every `(target, medium)` known for this recipient, each with the state a send would **actually** act on.\n\n`state.enabled` is the **resolved** decision — the same answer the delivery path computes — not a raw stored row. Resolution walks, in order:\n\n1. the recipient's rule for the exact target,\n2. the recipient's `topic: any` rule for the channel/event,\n3. the project's rule for the exact target,\n4. the project's `topic: any` rule,\n5. the medium's **default**.\n\n**The default is per medium:** `in_app` delivers, every other medium does not. A `topic: none` target never takes an `any` rule.\n\nThe response covers every target in the project catalog **plus** any target this recipient has a rule of their own for, across the active mediums (`in_app` and `email`). So a recipient can have an entry for a `(target, medium)` you never cataloged — if they have an explicit rule for it, it is live, and this reports it.\n\nEach `state` carries:\n\n-   `enabled` — whether a send to this `(target, medium)` would deliver.\n-   `inherited` — `true` when the recipient has no rule of their own for this exact `(target, medium)`; the value came from elsewhere in the cascade. `false` means they set it explicitly.\n-   `cataloged` — whether a project-level rule exists for this exact `(target, medium)`. This is **context, not a gate**: an explicit recipient rule on an uncataloged pair still delivers, and `in_app` delivers by default with no catalog entry at all. Use it to decide what to render, never to predict delivery — `enabled` is the answer.",
                "parameters": [
                    {
                        "name": "recipient_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The unique identifier of the recipient."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of recipient's resolved preferences.",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Preferences": {
                                        "summary": "Preferences Response",
                                        "value": {
                                            "data": {
                                                "preferences": [
                                                    {
                                                        "target": {
                                                            "channel": "posts",
                                                            "topic": "any",
                                                            "event": "new_comment",
                                                            "medium": "in_app",
                                                            "label": "Comments on your posts"
                                                        },
                                                        "state": {
                                                            "enabled": true,
                                                            "inherited": true,
                                                            "cataloged": true
                                                        }
                                                    },
                                                    {
                                                        "target": {
                                                            "channel": "announcements",
                                                            "topic": "none",
                                                            "event": "new_feature",
                                                            "medium": "email",
                                                            "label": "Product Announcements"
                                                        },
                                                        "state": {
                                                            "enabled": false,
                                                            "inherited": false,
                                                            "cataloged": true
                                                        }
                                                    },
                                                    {
                                                        "target": {
                                                            "channel": "announcements",
                                                            "topic": "none",
                                                            "event": "new_feature",
                                                            "medium": "in_app"
                                                        },
                                                        "state": {
                                                            "enabled": true,
                                                            "inherited": true,
                                                            "cataloged": false
                                                        }
                                                    }
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuthWithAPIKeyWithEitherScope": []
                    }
                ],
                "x-codeSamples": [
                    {
                        "lang": "curl",
                        "label": "cURL",
                        "source": "curl https://api.bodhveda.com/recipients/recipient_123/preferences \\\n  -H \"Authorization: Bearer bv_xxxxxxxxx\" "
                    },
                    {
                        "lang": "javascript",
                        "label": "JavaScript",
                        "source": "import { Bodhveda } from \"@bodhveda/js\";\n\nconst bodhveda = new Bodhveda(\"bv_xxxxxxxxx\");\n\nconst { preferences } = await bodhveda.recipients.preferences.list(\"recipient_123\");"
                    },
                    {
                        "lang": "go",
                        "label": "Go",
                        "source": "import (\n    \"context\"\n\n    bodhveda \"github.com/MudgalLabs/bodhveda/sdk/go\"\n)\n\nctx := context.Background()\nclient := bodhveda.NewClient(\"bv_xxxxxxxxx\", nil)\n\nres, _ := client.Recipients.Preferences.List(ctx, \"recipient_123\")\n// res.Preferences"
                    },
                    {
                        "lang": "jsx",
                        "label": "React",
                        "source": "import { usePreferences } from \"@bodhveda/react\";\n\n// Requires a <BodhvedaProvider recipientID=\"recipient_123\"> higher in the tree.\nfunction Preferences() {\n    const { data, isLoading } = usePreferences();\n    if (isLoading) return null;\n    return <PreferenceList items={data.preferences} />;\n}"
                    }
                ]
            },
            "patch": {
                "summary": "Set recipient's subscription state for a target",
                "operationId": "setRecipientPreference",
                "tags": [
                    "Preferences"
                ],
                "description": "Set a recipient's subscription state for a specific target.",
                "parameters": [
                    {
                        "name": "recipient_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The unique identifier of the recipient."
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "target": {
                                        "$ref": "#/components/schemas/Target"
                                    },
                                    "medium": {
                                        "type": "string",
                                        "description": "The medium this preference applies to. In-app and email are toggled independently for the same target. Defaults to `in_app` when omitted.",
                                        "enum": [
                                            "in_app",
                                            "email"
                                        ],
                                        "default": "in_app"
                                    },
                                    "state": {
                                        "type": "object",
                                        "properties": {
                                            "enabled": {
                                                "type": "boolean",
                                                "description": "Whether the recipient is subscribed to this target for this medium."
                                            }
                                        },
                                        "required": [
                                            "enabled"
                                        ]
                                    }
                                },
                                "required": [
                                    "target",
                                    "state"
                                ]
                            }
                        }
                    }
                },
                "x-codeSamples": [
                    {
                        "lang": "curl",
                        "label": "cURL",
                        "source": "curl -X PATCH https://api.bodhveda.com/recipients/recipient_123/preferences \\\n  -H \"Authorization: Bearer bv_xxxxxxxxx\" \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\n    \"target\": {\n      \"channel\": \"posts\",\n      \"topic\": \"post_id_123\",\n      \"event\": \"new_comment\"\n    },\n    \"medium\": \"email\",\n    \"state\": {\n      \"enabled\": false\n    }\n  }' "
                    },
                    {
                        "lang": "javascript",
                        "label": "JavaScript",
                        "source": "import { Bodhveda } from \"@bodhveda/js\";\n\nconst bodhveda = new Bodhveda(\"bv_xxxxxxxxx\");\n\nconst pref = await bodhveda.recipients.preferences.set(\"recipient_123\", {\n    target: { channel: \"posts\", topic: \"post_id_123\", event: \"new_comment\" },\n    medium: \"email\",\n    state: { enabled: false },\n});"
                    },
                    {
                        "lang": "go",
                        "label": "Go",
                        "source": "import (\n    \"context\"\n\n    bodhveda \"github.com/MudgalLabs/bodhveda/sdk/go\"\n)\n\nctx := context.Background()\nclient := bodhveda.NewClient(\"bv_xxxxxxxxx\", nil)\n\nreq := &bodhveda.SetPreferenceRequest{\n    Target: bodhveda.Target{Channel: \"posts\", Topic: \"post_id_123\", Event: \"new_comment\"},\n    Medium: bodhveda.MediumEmail,\n}\nreq.State.Enabled = false\npref, _ := client.Recipients.Preferences.Set(ctx, \"recipient_123\", req)"
                    },
                    {
                        "lang": "jsx",
                        "label": "React",
                        "source": "import { useUpdatePreference } from \"@bodhveda/react\";\n\n// Requires a <BodhvedaProvider recipientID=\"recipient_123\"> higher in the tree.\nfunction Toggle() {\n    const { mutate } = useUpdatePreference();\n    return (\n        <button\n            onClick={() =>\n                mutate({\n                    target: { channel: \"posts\", topic: \"post_id_123\", event: \"new_comment\" },\n                    medium: \"email\",\n                    state: { enabled: false },\n                })\n            }\n        >\n            Unsubscribe\n        </button>\n    );\n}"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated subscription state for the target",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Updated Target Subscription State": {
                                        "summary": "Updated Target Subscription State Response",
                                        "value": {
                                            "data": {
                                                "target": {
                                                    "channel": "posts",
                                                    "topic": "post_id_123",
                                                    "event": "new_comment",
                                                    "medium": "email"
                                                },
                                                "state": {
                                                    "enabled": false,
                                                    "inherited": false
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuthWithAPIKeyWithEitherScope": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuthWithAPIKeyWithFullAcessScope": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "APIKey",
                "description": "Bearer authentication header of the form `Bearer bv_xxxxxxxxx`, where `bv_xxxxxxxxx` is your **API Key** with `Full access` scope."
            },
            "BearerAuthWithAPIKeyWithEitherScope": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "APIKey",
                "description": "Bearer authentication header of the form `Bearer bv_xxxxxxxxx`, where `bv_xxxxxxxxx` is your **API Key** with `Full access` or `Recipient acess` scope."
            }
        },
        "schemas": {
...
            "Target": {
                "type": "object",
                "description": "A **target** is the structure used to describe _who should get a notification_ and _how it should be categorized_.",
                "properties": {
                    "channel": {
                        "type": "string",
                        "description": "Broad category for a notification (e.g., \"posts\")."
                    },
                    "topic": {
                        "type": "string",
                        "description": "Subcategory inside the channel (e.g., \"post_id_123\"). \n- `any` and `none` have special meaning. \n- `any` means **all topics under this channel/event**. \n- `none` means **only the base channel/event**. \n- `any` is **not** a valid `topic` when sending a notification, but both are valid for preferences. \n\n A preference with the target `posts:any:new_comment` will match a notification with target `posts:post_id_123:new_comment` but a preference with the target `posts:none:new_comment` will **only** match a notification with target `posts:none:new_comment`."
                    },
                    "event": {
                        "type": "string",
                        "description": "Event that triggered this notification (e.g., \"new_comment\")."
                    }
                },
                "required": [
                    "channel",
                    "topic",
                    "event"
                ]
            },
...
        }
    }
}